Feature/metro api

+++ src/main/kotlin/ru/mosmetro/backend/model/dto/order/PassengerOrderDTO.kt
@@ -17,7 +17,7 @@ data class PassengerOrderDTO(
     @Schema(description = "Способ приема заявки")
     val orderApplication: OrderApplicationDTO,
     @Schema(description = "Время заявки в секундах")
-    val duration: Double,
+    val duration: Long,
     @Schema(description = "Количество пассажиров")
     val passengerCount: Int,
     @Schema(description = "Количество работников мужского пола, нужное для данной заявки")
